+++ docs/changelog.md
@@ -41,6 +41,9 @@ What's changed since v1.47.0:
   - Azure Container Registry:
     - Check that audit diagnostic logs are enabled for Container Registry by @BernieWhite.
       [#3445](https://github.com/Azure/PSRule.Rules.Azure/issues/3445)
+  - Automation Account:
+    - Added `Azure.Automation.Runbook.NotPinnedUri` to check that runbooks use pinned script dependencies when sourcing from `raw.githubusercontent.com`.
+      [#3324](https://github.com/Azure/PSRule.Rules.Azure/issues/3324)
   - Container Apps:
     - Check that liveness and readiness health probes use HTTP checks for HTTP-based ingress by @BernieWhite.
       [#3111](https://github.com/Azure/PSRule.Rules.Azure/issues/3111)